// src/app/body/report.interface.ts
export interface Report {
  blsEmploymentMay2018: number;
  blsEmploymentPerThousandMay2018: number;
  blsMeanWageMay2018: number;
  day: number;
  glassdoorJobs: number;
  glassdoorSalary: number;
  glassdoorSalarySamples: number;
  indeedJobs: number;
  indeedSalary: number;
  indeedSalarySamples: number;
  title: string;
  metro: string;
  month: number;
  year: number;
}